+++ app/src/main/java/com/avon/spott/PhotoMapFragment.kt
@@ -6,12 +6,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import kotlinx.android.synthetic.main.toolbar.*
 
-class HomeFragment : Fragment() {
-
+class PhotoMapFragment: Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val root = inflater.inflate(R.layout.fragment_home, container, false)
+        val root = inflater.inflate(R.layout.fragment_photomap, container, false)
         return root
+
     }
 }
