+++ app/src/main/java/com/avon/spott/ScrapFragment.kt
@@ -7,6 +7,6 @@ class LikeFragment : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.fragment_like)
+        setContentView(R.layout.fragment_scrap)
     }
 }
