--- app/src/main/res/drawable-hdpi/google_login_button.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-<item>
-    <shape>
-        <stroke android:color="@color/colorGoogle"/>
-        <solid android:color="@color/colorGoogle"/>
-        <corners android:radius="4dp"/>
-    </shape>
-</item>
-</selector>
