--- app/src/main/res/layout/activity_mypage.xml
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MypageActivity">
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/floatingbtn_addphoto_mypage"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="20dp"
-        android:layout_marginRight="20dp"
-        android:layout_marginBottom="20dp"
-        android:clickable="true"
-        android:src="@drawable/ic_add_photo_alternate_24px"
-        app:backgroundTint="@color/colorPrimary"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        tools:ignore="VectorDrawableCompat" />
-
-    <include
-        android:id="@+id/include_toolbar_mypage_f"
-        layout="@layout/toolbar"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="40dp"
-        android:orientation="horizontal"
-        android:id="@+id/tab_photomap_f"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/include_toolbar_mypage_f"
-        app:tabIndicatorColor="@color/colorPrimary">
-
-        <com.google.android.material.tabs.TabItem
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:icon="@drawable/ic_grid_on_primary_24dp"/>
-
-        <com.google.android.material.tabs.TabItem
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:icon="@drawable/ic_map_primary_24dp"/>
-
-
-<!--    <androidx.viewpager.widget.ViewPager-->
-<!--        android:id="@+id/container_photomap_f"-->
-<!--        android:layout_width="0dp"-->
-<!--        android:layout_height="0dp"-->
-<!--        app:layout_constraintBottom_toBottomOf="parent"-->
-
-<!--        app:layout_constraintEnd_toEndOf="parent"-->
-<!--        app:layout_constraintStart_toStartOf="parent"-->
-<!--        app:layout_constraintTop_toBottomOf="@+id/include_toolbar_mypage_f"/>-->
-
-<!--        <ImageButton-->
-<!--            android:id="@+id/imgbtn_grid_mypage_f"-->
-<!--            android:layout_width="fill_parent"-->
-<!--            android:layout_height="fill_parent"-->
-<!--            android:background="@drawable/btn_left_mypage_f"-->
-<!--            app:srcCompat="@drawable/ic_grid_on_selector"-->
-<!--            android:layout_weight="1"/>-->
-
-<!--        <ImageButton-->
-<!--            android:id="@+id/imgbtn_map_mypage_f"-->
-<!--            android:layout_width="fill_parent"-->
-<!--            android:layout_height="fill_parent"-->
-<!--            android:background="@drawable/btn_right_mypage_f"-->
-<!--            app:srcCompat="@drawable/ic_map_selector"-->
-<!--            android:layout_weight="1"/>-->
-
-    </LinearLayout>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
