+++ app/src/main/res/layout/fragment_photo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PhotoActivity">
+    tools:context=".PhotoFragment">
 
     <include
         android:id="@+id/include_toolbar_photo_f"
