+++ app/src/main/res/layout/fragment_scrap.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LikeFragment">
+    tools:context=".ScrapFragment">
 
     <TextView
         android:id="@+id/text_count_like_f"
