+++ app/src/main/res/values/colors.xml
@@ -7,6 +7,7 @@
     <!-- 화면 배경색 -->
     <color name="default_background">#FAFAFA</color>
     <color name="bg_black">#050505</color>
+    <color name="status_bar">#22272B</color>
 
     <!-- [임시] 배경회색 -->
     <color name="background_grey">#B8B8B8</color>
